fix disount

+++ src/main/java/com/example/shoppingweb/model/Discount.java
@@ -21,5 +21,5 @@ public class Discount {
     private String discountName;
 
     @Column
-    private int discountPrice;
+    private Double discountPrice;
 }
